pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-agent-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
]
